Added card: Ashen | Updated destroyMinion function
destroyMinion function now does more things related to the destruction of a minion. It is called by the monsterInfo script as well. Another update: a minion that does not have an On Summon effect will not trigger an On Summon function. So, no reason for implementation of empty On Summon functions.